// web-ui/packages/app/lib/ai/tools/types.ts
export type CaseFileSearchOptions = HybridSearchOptions & {
  scope?: CaseFileSearchScopeType[];
  emailId?: string; // email_id
  threadId?: string; // thread_id
  attachmentId?: number; // attachment_id
  documentId?: number; // id
  replyToDocumentId?: number; // parent_email_id
  relatedToDocumentId?: number; // relatedEmailIds
};